requirements.txt
aiogram~=2.12.1
tortoise-orm~=0.17.1